esp-server/pyproject.toml
[project]
name = "esp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "fastapi[standard]>=0.121.0",
    "faster-whisper==1.2.1",
    "google-genai==1.48.0",
    "markdown>=3.10",
    "piper-tts==1.3.0",
    "setuptools==66.1.1",
    "pygame>=2.6.1",
    "uvicorn>=0.38.0",
    "watchdog>=6.0.0",
    "sounddevice>=0.5.3",
    "websockets>=15.0.1",
    "soundfile>=0.13.1",
    "numpy<2.2",
    "scipy>=1.16.3",
    "librosa>=0.11.0",
    "ipykernel>=7.1.0",
]
devDependencies = [
]
